feat: temporarily remove links to the registered words
It does not work well with Nuxt's SSG for now since the links returns HTTP 302. (redirects to the localized pages.)
To avoid that, you have to replace the links with the paths with language code (e.g. "/gold" -> "/en/gold"), but it is a bit hard to implement that with current implementation, so I decided to remove the links for now.